udpate

+++ db/migration/000001_init_schema.up.sql
@@ -6,7 +6,7 @@ CREATE TABLE "accounts" (
   "created_at" timestamp NOT NULL DEFAULT (now())
 );
 
-CREATE TABLE "entries" ("id" bigserial PRIMARY KEY, "account_id" bigint NOT NULL, "amount" NUMERIC (15, 2) NOT NULL, "created_at" timestamp NOT NULL DEFAULT (now()));
+CREATE TABLE "entries" ("id" bigserial PRIMARY KEY, "account_id" bigint NOT NULL, "amount" bigint NOT NULL, "created_at" timestamp NOT NULL DEFAULT (now()));
 
 CREATE TABLE "transfers" (
   "id" bigserial PRIMARY KEY,
